+++ src/custom-snippets.ts
@@ -1,10 +1,13 @@
 import * as vscode from 'vscode';
 import * as _ from 'lodash';
-import * as state from './state';
 import * as util from './utilities';
+import * as getText from './util/get-text';
 import * as namespace from './namespace';
 import * as outputWindow from './results-output/results-doc'
-import { customREPLCommandSnippet, evaluateInOutputWindow } from './evaluate';
+import { customREPLCommandSnippet } from './evaluate';
+import { getConfig } from './config';
+import * as replSession from './nrepl/repl-session';
+import * as evaluate from './evaluate';
 
 export async function evaluateCustomCodeSnippetCommand(codeOrKey?: string) {
     await evaluateCustomCodeSnippet(codeOrKey);
@@ -17,15 +20,15 @@ async function evaluateCustomCodeSnippet(codeOrKey?: string): Promise<void> {
     const currentFilename = editor.document.fileName;
     let pickCounter = 0;
     let configErrors: { "name": string; "keys": string[]; }[] = [];
-    const globalSnippets = state.config().customREPLCommandSnippetsGlobal as customREPLCommandSnippet[];
-    const workspaceSnippets = state.config().customREPLCommandSnippetsWorkspace as customREPLCommandSnippet[];
-    const workspaceFolderSnippets = state.config().customREPLCommandSnippetsWorkspaceFolder as customREPLCommandSnippet[];
+    const globalSnippets = getConfig().customREPLCommandSnippetsGlobal as customREPLCommandSnippet[];
+    const workspaceSnippets = getConfig().customREPLCommandSnippetsWorkspace as customREPLCommandSnippet[];
+    const workspaceFolderSnippets = getConfig().customREPLCommandSnippetsWorkspaceFolder as customREPLCommandSnippet[];
     let snippets = [
         ...(globalSnippets ? globalSnippets : []),
         ...(workspaceSnippets ? workspaceSnippets : []),
         ...(workspaceFolderSnippets ? workspaceFolderSnippets : [])];
     if (snippets.length < 1) {
-        snippets = state.config().customREPLCommandSnippets;
+        snippets = getConfig().customREPLCommandSnippets;
     }
     const snippetsDict = {};
     const snippetsKeyDict = {};
@@ -33,7 +36,7 @@ async function evaluateCustomCodeSnippet(codeOrKey?: string): Promise<void> {
     const editorNS = editor && editor.document &&
         editor.document.languageId === 'clojure' ? namespace.getNamespace(editor.document) : undefined;
     const editorRepl = editor && editor.document &&
-        editor.document.languageId === 'clojure' ? namespace.getREPLSessionType() : "clj";
+        editor.document.languageId === 'clojure' ? replSession.getReplSessionTypeFromState() : "clj";
     snippets.forEach((c: customREPLCommandSnippet) => {
         const undefs = ["name", "snippet"].filter(k => {
             return !c[k];
@@ -89,10 +92,13 @@ async function evaluateCustomCodeSnippet(codeOrKey?: string): Promise<void> {
         replace("$column", currentColumn).
         replace("$file", currentFilename).
         replace("$ns", ns).
-        replace("$current-form", util.currentFormText(editor, false)).
-        replace("$top-level-form", util.currentFormText(editor, true)).
-        replace("$current-fn", util.currentFunction(editor)).
-        replace("$top-level-defined-symbol", util.currentTopLevelFunction(editor));
-    await evaluateInOutputWindow(interpolatedCode, repl, ns);
+        replace("$selection", editor.document.getText(editor.selection)).
+        replace("$current-form", getText.currentFormText(editor)[1]).
+        replace("$top-level-form", getText.currentTopLevelFormText(editor)[1]).
+        replace("$current-fn", getText.currentFunction(editor)[1]).
+        replace("$top-level-defined-symbol", getText.currentTopLevelFunction(editor)[1]).
+        replace("$head", getText.toStartOfList(editor)[1]).
+        replace("$tail", getText.toEndOfList(editor)[1]);
+    await evaluate.evaluateInOutputWindow(interpolatedCode, repl, ns);
     outputWindow.appendPrompt();
 }