+++ src/nrepl/index.ts
@@ -2,7 +2,6 @@ import * as net from "net";
 import { BEncoderStream, BDecoderStream } from "./bencode";
 import * as state from './../state';
 import * as util from '../utilities';
-import { prettyPrint } from '../../out/cljs-lib/cljs-lib';
 import { PrettyPrintingOptions, disabledPrettyPrinter, getServerSidePrinter } from "../printer";
 import * as debug from "../debugger/calva-debug";
 import * as vscode from 'vscode';
@@ -10,6 +9,8 @@ import debugDecorations from '../debugger/decorations';
 import * as outputWindow from '../results-output/results-doc';
 import { formatAsLineComments } from '../results-output/util';
 import type { ReplSessionType } from '../config';
+import { getStateValue, prettyPrint } from '../../out/cljs-lib/cljs-lib';
+import { getConfig } from '../config';
 
 /** An nREPL client */
 export class NReplClient {
@@ -264,7 +265,7 @@ export class NReplSession {
 
     private _createEvalOperationMessage(code: string, ns: string, opts: any) {
         if (vscode.debug.activeDebugSession && this.replType === 'clj') {
-            const debugResponse = state.deref().get(debug.DEBUG_RESPONSE_KEY);
+            const debugResponse = getStateValue(debug.DEBUG_RESPONSE_KEY);
             state.analytics().logEvent(debug.DEBUG_ANALYTICS.CATEGORY, debug.DEBUG_ANALYTICS.EVENT_ACTIONS.EVALUATE_IN_DEBUG_CONTEXT).send();
             return {
                 id: debugResponse.id,
@@ -359,7 +360,7 @@ export class NReplSession {
     complete(ns: string, symbol: string, context?: string) {
         return new Promise<any>((resolve, reject) => {
             const id = this.client.nextId,
-                extraOpts = state.config().enableJSCompletions ? { "enhanced-cljs-completion?": "t" } : {};
+                extraOpts = getConfig().enableJSCompletions ? { "enhanced-cljs-completion?": "t" } : {};
             this.messageHandlers[id] = (msg) => {
                 resolve(msg);
                 return true;