+++ src/providers/completion.ts
@@ -5,6 +5,7 @@ import select from '../select';
 import * as docMirror from '../doc-mirror/index';
 import * as infoparser from './infoparser';
 import * as namespace from '../namespace';
+import * as replSession from '../nrepl/repl-session';
 
 export default class CalvaCompletionItemProvider implements CompletionItemProvider {
     state: any;
@@ -40,7 +41,7 @@ export default class CalvaCompletionItemProvider implements CompletionItemProvid
                 context = `${contextStart}__prefix__${contextEnd}`,
                 toplevelIsValidForm = toplevelStartCursor.withinValidList() && context != '__prefix__',
                 ns = namespace.getNamespace(document),
-                client = namespace.getSession(util.getFileType(document)),
+                client = replSession.getSession(util.getFileType(document)),
                 res = await client.complete(ns, text, toplevelIsValidForm ? context : null),
                 results = res.completions || [];
                 if(results) {
@@ -63,13 +64,13 @@ export default class CalvaCompletionItemProvider implements CompletionItemProvid
     }
 
     async resolveCompletionItem(item: CompletionItem, token: CancellationToken) {
-
         if (util.getConnectedState()) {
-            let client = namespace.getSession(util.getFileType(window.activeTextEditor.document));
+            let client = replSession.getSession(util.getFileType(window.activeTextEditor.document));
             if (client) {
                 await namespace.createNamespaceFromDocumentIfNotExists(window.activeTextEditor.document);
-                let result = await client.info(item.insertText["ns"], item.label)
-                let [doc, details] = infoparser.getCompletion(result);
+                const ns = namespace.getDocumentNamespace();
+                const result = await client.info(ns, item.label)
+                const [doc, details] = infoparser.getCompletion(result);
                 item.documentation = doc;
                 item.detail = details;
             }