+++ src/results-output/results-doc.ts
@@ -1,6 +1,5 @@
 import * as path from 'path';
 import * as vscode from 'vscode';
-import * as os from 'os';
 import * as state from '../state';
 import { highlight } from '../highlight/src/extension'
 import { NReplSession } from '../nrepl';
@@ -8,11 +7,12 @@ import * as util from '../utilities';
 import select from '../select';
 import { formatCode } from '../calva-fmt/src/format';
 import * as namespace from '../namespace';
-import config from '../config';
+import * as config from '../config';
 import type { ReplSessionType } from '../config';
 import * as replHistory from './repl-history';
 import * as docMirror from '../doc-mirror/index'
 import { PrintStackTraceCodelensProvider } from '../providers/codelense';
+import * as replSession from '../nrepl/repl-session';
 
 const RESULTS_DOC_NAME = `output.${config.REPL_FILE_EXT}`;
 
@@ -126,7 +126,7 @@ export async function initResultsDoc(): Promise<vscode.TextDocument> {
     await vscode.workspace.applyEdit(edit);
     resultsDoc.save();
 
-    if (state.config().autoOpenREPLWindow) {
+    if (config.getConfig().autoOpenREPLWindow) {
         const resultsEditor = await vscode.window.showTextDocument(resultsDoc, getViewColumn(), true);
         const firstPos = resultsEditor.document.positionAt(0);
         const lastPos = resultsDoc.positionAt(Infinity);
@@ -163,7 +163,7 @@ export async function initResultsDoc(): Promise<vscode.TextDocument> {
         }
         vscode.commands.executeCommand("setContext", "calva:outputWindowSubmitOnEnter", submitOnEnter);
     }));
-    vscode.languages.registerCodeLensProvider(state.documentSelector, new PrintStackTraceCodelensProvider());
+    vscode.languages.registerCodeLensProvider(config.documentSelector, new PrintStackTraceCodelensProvider());
 
     // If the output window is active when initResultsDoc is run, these contexts won't be set properly without the below
     // until the next time it's focused
@@ -194,17 +194,17 @@ export async function revealDocForCurrentNS(preserveFocus: boolean = true) {
 }
 
 export async function setNamespaceFromCurrentFile() {
-    const session = namespace.getSession();
+    const session = replSession.getSession();
     const ns = namespace.getNamespace(util.getDocument({}));
     if (getNs() !== ns) {
         await session.eval("(in-ns '" + ns + ")", session.client.ns).value;
     }
     setSession(session, ns);
-    namespace.updateREPLSessionType();
+    replSession.updateReplSessionType();
 }
 
 async function appendFormGrabbingSessionAndNS(topLevel: boolean) {
-    const session = namespace.getSession();
+    const session = replSession.getSession();
     const ns = namespace.getNamespace(util.getDocument({}));
     const editor = vscode.window.activeTextEditor;
     const doc = editor.document;
@@ -237,7 +237,7 @@ let scrollToBottomSub: vscode.Disposable;
 export interface OnAppendedCallback {
     (insertLocation: vscode.Location, newPosition?: vscode.Location): any
 }
-const editQueue: [string, OnAppendedCallback][] = [];
+let editQueue: [string, OnAppendedCallback][] = [];
 let applyingEdit = false;
 /* Because this function can be called several times asynchronously by the handling of incoming nrepl messages and those,
    we should never await it, because that await could possibly not return until way later, after edits that came in from elsewhere
@@ -301,6 +301,11 @@ export function append(text: string, onAppended?: OnAppendedCallback): void {
             }
         });
     };
+}
+
+export function discardPendingPrints(): void {
+    editQueue = [];
+    appendPrompt();
 }
 
 export type OutputStacktraceEntry = { uri: vscode.Uri, line: number };