+++ src/select.ts
@@ -2,12 +2,16 @@ import * as vscode from 'vscode';
 import * as util from './utilities'
 import * as docMirror from './doc-mirror/index'
 
+function selectionFromOffsetRange(doc: vscode.TextDocument, range: [number, number]): vscode.Selection {
+    return new vscode.Selection(doc.positionAt(range[0]), doc.positionAt(range[1]));
+}
+
 function getFormSelection(doc: vscode.TextDocument, pos: vscode.Position, topLevel): vscode.Selection {
     const idx = doc.offsetAt(pos),
         cursor = docMirror.getDocument(doc).getTokenCursor(topLevel ? 0 : idx),
         range = topLevel ? cursor.rangeForDefun(idx) : cursor.rangeForCurrentForm(idx);
     if (range) {
-        return new vscode.Selection(doc.positionAt(range[0]), doc.positionAt(range[1]));
+        return selectionFromOffsetRange(doc, range);
     }
 }
 
@@ -26,5 +30,6 @@ function selectCurrentForm(document = {}) {
 
 export default {
     getFormSelection: getFormSelection,
-    selectCurrentForm: selectCurrentForm
+    selectCurrentForm: selectCurrentForm,
+    selectionFromOffsetRange
 };